+++ ignite/templates/typed/options.go
@@ -14,7 +14,6 @@ import (
 // Options ...
 type Options struct {
 	AppName      string
-	AppPath      string
 	ProtoDir     string
 	ProtoVer     string
 	ModuleName   string
@@ -30,7 +29,7 @@ type Options struct {
 
 // ProtoFile returns the path to the proto folder within the generated app.
 func (opts *Options) ProtoFile(fname string) string {
-	return filepath.Join(opts.AppPath, opts.ProtoDir, opts.AppName, opts.ModuleName, opts.ProtoVer, fname)
+	return filepath.Join(opts.ProtoDir, opts.AppName, opts.ModuleName, opts.ProtoVer, fname)
 }
 
 // ProtoTypeImport Return the protobuf import statement for this type.